requirements.txt
func_timeout==4.3.5
pycryptodomex==3.17
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
revChatGPT==4.0.0
web.py==0.61
